# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.10",
    "title": "META API",
    "description": "API文件"
  },
  "host": "blog-server-izfg.onrender.com",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "jwt token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/example/": {
      "get": {
        "tags": [
          "Example"
        ],
        "description": "取得範例資料",
        "parameters": [],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "string",
                  "example": "Hello World"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/posts/": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "取得貼文列表",
        "description": "取得貼文資料",
        "parameters": [
          {
            "name": "timeSort",
            "in": "query",
            "description": "排序方式 asc/desc",
            "type": "string"
          },
          {
            "name": "q",
            "in": "query",
            "description": "模糊搜尋",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳貼文資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "642bda33a4ab589d9a9cc291"
                    },
                    "content": {
                      "type": "string",
                      "example": "測試"
                    },
                    "image": {
                      "type": "string",
                      "example": ""
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "642bd0474819e89ecb878c4a"
                        },
                        "name": {
                          "type": "string",
                          "example": "John"
                        },
                        "photo": {
                          "type": "string",
                          "example": "https://thumb.fakeface.rest/thumb_male_10_8c02e4e9bdc0e103530691acfca605f18caf1766.jpg"
                        }
                      }
                    },
                    "likes": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "summary": "新增貼文",
        "description": "新增貼文資料",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add a user",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "測試3"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "string",
                      "example": "測試3"
                    },
                    "image": {
                      "type": "string",
                      "example": ""
                    },
                    "createAt": {
                      "type": "string",
                      "example": "2023-04-04T08:10:25.698Z"
                    },
                    "user": {
                      "type": "string",
                      "example": "642bd0474819e89ecb878c4a"
                    },
                    "likes": {
                      "type": "number",
                      "example": 0
                    },
                    "_id": {
                      "type": "string",
                      "example": "642bdba2712ab795131db4e4"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/users/sign_up": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "使用者註冊",
        "description": "使用者註冊",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "註冊資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "jason"
                },
                "email": {
                  "type": "string",
                  "example": "jason@gmail.com"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "jason"
                },
                "password": {
                  "type": "string",
                  "example": "jason"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳註冊資訊與token",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "jason"
                        },
                        "email": {
                          "type": "string",
                          "example": "z27089433@gmail.com"
                        },
                        "photo": {
                          "type": "string",
                          "example": ""
                        },
                        "password": {
                          "type": "string",
                          "example": "$2a$12$1/vlSkyJDXkJFkgUlHCqPOMlUaW3mXibBOj/mcH0AKCKzm42KuULa"
                        },
                        "_id": {
                          "type": "string",
                          "example": "643b854fadd06971aca8a0df"
                        }
                      }
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0M2I4NTRmYWRkMDY5NzFhY2E4YTBkZiIsImlhdCI6MTY4MTYyODc2NywiZXhwIjoxNjg0MjIwNzY3fQ.jgzymQFGUEraIaLDl4dOPRvaPb_H3yZLYZ4JVMrrrPM"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2ODE2MjgyNDAsImV4cCI6MTY4NDIyMDI0MH0.gwSnMNRmu20n4OWW_-_oEI81RjWH_9ZB5RwvPP2cHbE"
          }
        }
      }
    },
    "/api/users/sign_in": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "使用者登入",
        "description": "使用者登入",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "登入帳密",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "jason@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "jason"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳登入者訊息",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "643b854fadd06971aca8a0df"
                        },
                        "user": {
                          "type": "string",
                          "example": "jason"
                        },
                        "email": {
                          "type": "string",
                          "example": "z27089433@gmail.com"
                        },
                        "photo": {
                          "type": "string",
                          "example": "123"
                        }
                      }
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0M2I4NTRmYWRkMDY5NzFhY2E4YTBkZiIsImlhdCI6MTY4MTYyODc2NywiZXhwIjoxNjg0MjIwNzY3fQ.jgzymQFGUEraIaLDl4dOPRvaPb_H3yZLYZ4JVMrrrPM"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2ODE2MjgyNDAsImV4cCI6MTY4NDIyMDI0MH0.gwSnMNRmu20n4OWW_-_oEI81RjWH_9ZB5RwvPP2cHbE"
          }
        }
      }
    },
    "/api/users/update_password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "更新使用者密碼",
        "description": "更新使用者密碼",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "使用者新舊帳密與email",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "jason@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "jason"
                },
                "passwordNew": {
                  "type": "string",
                  "example": "jason"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳更新成功與否",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "isUpdate": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/users/profile": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "更新登入者訊息",
        "description": "更新登入者訊息",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "登入者訊息",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "jason"
                },
                "photo": {
                  "type": "string",
                  "example": "jason"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳登入訊息與token",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "643b854fadd06971aca8a0df"
                        },
                        "email": {
                          "type": "string",
                          "example": "jason@gmail.com"
                        },
                        "name": {
                          "type": "string",
                          "example": "jason"
                        },
                        "photo": {
                          "type": "string",
                          "example": "123"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2ODE2MjgyNDAsImV4cCI6MTY4NDIyMDI0MH0.gwSnMNRmu20n4OWW_-_oEI81RjWH_9ZB5RwvPP2cHbE"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "登入者訊息",
        "description": "登入者訊息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "回傳登入者訊息",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "643b854fadd06971aca8a0df"
                        },
                        "name": {
                          "type": "string",
                          "example": "jason"
                        },
                        "email": {
                          "type": "string",
                          "example": "z27089433@gmail.com"
                        },
                        "photo": {
                          "type": "string",
                          "example": "123"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {}
}